+++ restart.sh
@@ -1,0 +1,1 @@
+clear && pip uninstall pySorts && pip cache purge && pip install . --verbose && ./up.sh
