--- setup.cfg
@@ -1,21 +1,0 @@
-[metadata]
-name = pySorts
-version = 0.0.20
-description = An application that compares sorting algorithm implementations by language and algorithm.
-author = Jon Atkinson
-license = MIT
-
-[options]
-package_dir =
-    = src
-packages = find:
-
-[options.entry_points]
-console_scripts =
-    pySortsApp = python.gui.main:main
-
-[options.packages.find]
-where = src
-
-[options.package_data]
-python = ../c/cSorts.so
