--- src/python/backend/setup.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-
-from distutils.core import setup
-
-setup(name='pySorts',
-      version='0.01',
-      description='simple sorts comparison',
-      author='Jon Atkinson'
-     )
